requirements.txt
psutil~=6.1.1
scapy~=2.6.1
python-nmap~=0.7.1
pythonping~=1.1.4
